+++ tsconfig.json
@@ -12,7 +12,7 @@
       "esModuleInterop": true,
     },
     "include": [
-      "src/**/*.ts"
+      "google-storage/src/**/*.ts"
     ],
     "exclude": [
       "node_modules"
